requirements.txt
requests
python-dotenv
pandas
pyarrow
scikit-learn
joblib
fastapi
uvicorn
mlflow
lightgbm
openai